requirements.txt
alabaster==0.7.12
arrow==1.2.3
astroid==2.12.13
asttokens==2.2.1
atomicwrites==1.4.1
attrs==22.2.0
autopep8==1.6.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
black==22.12.0
bleach==5.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
comm==0.1.2
cookiecutter==2.1.1
cryptography==38.0.4
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.6
docstring-to-markdown==0.11
docutils==0.19
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
flake8==6.0.0
idna==3.4
imagesize==1.4.1
importlib-metadata==5.2.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
isort==5.11.4
jaraco.classes==3.2.3
jedi==0.18.2
jeepney==0.8.0
jellyfish==0.9.0
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyterlab-pygments==0.2.2
keyring==23.13.1
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
more-itertools==9.0.0
mypy-extensions==0.4.3
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
numpydoc==1.5.0
packaging==22.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
pluggy==1.0.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==3.0.1
Pygments==2.13.0
PyJWT==2.6.0
pylint==2.15.9
pylint-venv==2.3.0
pyls-spyder==0.4.0
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-lsp-black==1.2.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.0
python-slugify==7.0.0
pytoolconfig==1.2.4
pytz==2022.7
pyxdg==0.28
PyYAML==6.0
pyzmq==24.0.1
QDarkStyle==3.0.3
qstylizer==0.2.2
QtAwesome==1.2.2
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
rope==1.6.0
Rtree==1.0.1
SecretStorage==3.3.3
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
Sphinx==6.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder-kernels==2.4.1
stack-data==0.6.2
text-unidecode==1.3
textdistance==4.5.0
three-merge==0.1.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
traitlets==5.8.0
twilio==7.16.0
ujson==5.6.0
urllib3==1.26.13
watchdog==2.2.0
wcwidth==0.2.5
webencodings==0.5.1
whatthepatch==1.0.3
wrapt==1.14.1
wurlitzer==3.0.3
yapf==0.32.0
zipp==3.11.0
